modules/book-updater/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "book-updater"
version = "0.1.0"
description = "Update Parasol V5 book content from latest commands and guides"
authors = [
    { name = "Parasol Team", email = "team@parasol.dev" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pathlib>=1.0.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
    "deepdiff>=6.0.0",
    "jinja2>=3.1.0",
    "gitpython>=3.1.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
bookupdater = "amplifier_module_book_updater.main:cli"

[tool.hatch.build]
include = [
    "amplifier_module_book_updater/**/*.py",
    "amplifier_module_book_updater/templates/**/*",
    "amplifier_module_book_updater/rules/**/*.yaml",
]